requirements.txt
streamlit==0.88.0
numpy==1.19.5
opencv-python==4.5.3.56
Pillow==8.3.2
deepface==0.0.68
